refactor(chat): extract reusable ChatPane
